# build: use faster minifier for js and css