--- a/webpack.config.ts
+++ b/webpack.config.ts
@@ -5,11 +5,14 @@ import fs from 'fs';
 import path from 'path';
 
 import {WebpackReactSourcemapsPlugin} from '@acemarke/react-prod-sourcemaps';
+import browserslist from 'browserslist';
 import CompressionPlugin from 'compression-webpack-plugin';
 import CopyPlugin from 'copy-webpack-plugin';
 import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
+import lightningcss from 'lightningcss';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import TerserPlugin from 'terser-webpack-plugin';
 import webpack from 'webpack';
 import {Configuration as DevServerConfig} from 'webpack-dev-server';
 import WebpackHookPlugin from 'webpack-hook-plugin';
@@ -20,6 +23,11 @@ import LastBuiltPlugin from './build-utils/last-built-plugin';
 import SentryInstrumentation from './build-utils/sentry-instrumentation';
 import {extractIOSDeviceNames} from './scripts/extract-ios-device-names';
 import babelConfig from './babel.config';
+import packageJson from './package.json';
+
+type MinimizerPluginOptions = {
+  targets: lightningcss.TransformAttributeOptions['targets'];
+};
 
 // Runs as part of prebuild step to generate a list of identifier -> name mappings for  iOS
 (async () => {
@@ -175,7 +183,7 @@ type CacheGroupTest = (
 ) => boolean;
 
 // A mapping of chunk groups used for locale code splitting
-const localeChunkGroups: CacheGroups = {};
+const cacheGroups: CacheGroups = {};
 
 supportedLocales
   // No need to split the english locale out as it will be completely empty and
@@ -207,7 +215,7 @@ supportedLocales
     //
     // In the application code you will still need to import via their module
     // paths and not the chunk name
-    localeChunkGroups[group] = {
+    cacheGroups[group] = {
       chunks: 'async',
       name: group,
       test: groupTest,
@@ -449,6 +457,7 @@ const appConfig: Configuration = {
 
     modules: ['node_modules'],
     extensions: ['.jsx', '.js', '.json', '.ts', '.tsx', '.less'],
+    symlinks: false,
   },
   output: {
     crossOriginLoading: 'anonymous',
@@ -470,14 +479,24 @@ const appConfig: Configuration = {
       chunks: 'async',
       maxInitialRequests: 10, // (default: 30)
       maxAsyncRequests: 10, // (default: 30)
-      cacheGroups: {
-        ...localeChunkGroups,
-      },
+      cacheGroups,
     },
 
-    // This only runs in production mode
-    // Grabbed this example from https://github.com/webpack-contrib/css-minimizer-webpack-plugin
-    minimizer: ['...', new CssMinimizerPlugin()],
+    minimizer: [
+      new TerserPlugin({
+        parallel: true,
+        minify: TerserPlugin.esbuildMinify,
+      }),
+      new CssMinimizerPlugin<MinimizerPluginOptions>({
+        parallel: true,
+        minify: CssMinimizerPlugin.lightningCssMinify,
+        minimizerOptions: {
+          targets: lightningcss.browserslistToTargets(
+            browserslist(packageJson.browserslist.production)
+          ),
+        },
+      }),
+    ],
   },
   devtool: IS_PRODUCTION ? 'source-map' : 'eval-cheap-module-source-map',
 };
@@ -753,15 +772,10 @@ const minificationPlugins = [
   // This compression-webpack-plugin generates pre-compressed files
   // ending in .gz, to be picked up and served by our internal static media
   // server as well as nginx when paired with the gzip_static module.
-  //
-  // TODO(ts): The current @types/compression-webpack-plugin is still targeting
-  //           webpack@4, for now we just as any it.
   new CompressionPlugin({
     algorithm: 'gzip',
     test: /\.(js|map|css|svg|html|txt|ico|eot|ttf)$/,
-  }) as any,
-  // NOTE: In production mode webpack will automatically minify javascript
-  // using the TerserWebpackPlugin.
+  }),
 ];
 
 if (IS_PRODUCTION) {