Uses esbuild and lightningCSS to minify JS and CSS.

Running `yarn build-production` command:

Before:

```
webpack 5.87.0 compiled with 1 warning in 33069 ms
✨  Done in 34.68s.
```

After:

```
webpack 5.87.0 compiled with 16 warnings in 22640 ms
✨  Done in 24.21s.
```