@ryan953 2024-01-04
in the description there is the Before speed and After speed. Can you write down the command that was executed so archaeologists know what to run.