@anonrig 2024-01-08
@getsentry/app-frontend Appreciate a review on this.